feat(platform): add template-based tab content for IconTabBar component
