+++ libs/docs/platform/icon-tab-bar/examples/platform-icon-tab-bar-template-content-example/platform-icon-tab-bar-template-content-example.component.html
@@ -9,6 +9,7 @@
 <ng-template fdpIconTabBarContentTemplate templateId="suppliers-template">
     <div>
         <h3>Supplier Directory</h3>
+        <p>Browse and manage your supplier relationships.</p>
     </div>
 </ng-template>
