+++ libs/platform/icon-tab-bar/directives/icon-tab-bar-content-template.directive.ts
@@ -1,37 +1,5 @@
-import { Directive, inject, Injectable, input, OnDestroy, OnInit, signal, Signal, TemplateRef } from '@angular/core';
-
-/**
- * Service to manage template registration across components
- */
-@Injectable({ providedIn: 'root' })
-export class IconTabBarTemplateService {
-    private templates = new Map<string, TemplateRef<any>>();
-    private templatesSignal = signal<Map<string, TemplateRef<any>>>(new Map());
-
-    registerTemplate(id: string, template: TemplateRef<any>): void {
-        this.templates.set(id, template);
-        // Create a new map to trigger signal update
-        this.templatesSignal.set(new Map(this.templates));
-    }
-
-    unregisterTemplate(id: string): void {
-        this.templates.delete(id);
-        // Create a new map to trigger signal update
-        this.templatesSignal.set(new Map(this.templates));
-    }
-
-    getTemplate(id: string): TemplateRef<any> | undefined {
-        return this.templates.get(id);
-    }
-
-    getAllTemplates(): Map<string, TemplateRef<any>> {
-        return new Map(this.templates);
-    }
-
-    getAllTemplatesSignal(): Signal<Map<string, TemplateRef<any>>> {
-        return this.templatesSignal.asReadonly();
-    }
-}
+import { Directive, inject, input, OnDestroy, OnInit, TemplateRef } from '@angular/core';
+import { IconTabBarTemplateService } from '../services/icon-tab-bar-template.service';
 
 /**
  * Directive to mark a template as a content template for icon tab bar.
