+++ libs/platform/icon-tab-bar/index.ts
@@ -12,6 +12,7 @@ export * from './icon-tab-bar.module';
 export * from './interfaces/icon-tab-bar-item.interface';
 export * from './interfaces/tab-color-associations.interface';
 export * from './interfaces/tab-config.interface';
+export * from './services/icon-tab-bar-template.service';
 export * from './types';
 
 export * from './components/text-type-tab-item/icon-tab-bar-text-type-tab-item.component';
